backend/requirements.txt
django>=5.2,<5.3
django-ninja>=1.1,<1.2
uvicorn[standard]>=0.30
psycopg[binary,pool]>=3.1,<4.0
dj-database-url>=2.1,<2.2
python-dotenv>=1.0,<1.1
django-cors-headers>=4.4
whitenoise>=6.6.0

channels>=4.1
channels-redis>=4.2

celery[redis]>=5.3,<5.4
redis>=5.0

pgvector>=0.2.4
requests>=2.32
